+++ CHANGES.rst
@@ -19,6 +19,10 @@ Changelog
   `OwnOrganizationServicesVocabulary` so methods `_term_value` and
   `_term_token` are available.
   [gbastien]
+- In `@@display-group-users` instead displaying contained groups without
+  contained members, display members of contained groups for normal users,
+  display contained groups and contained members to Managers.
+  [gbastien]
 
 1.29.1 (2020-08-26)
 -------------------